+++ src/main/java/com/example/Coin/services/coin/CoinService.java
@@ -2,13 +2,12 @@ package com.example.Coin.services.coin;
 
 public interface CoinService {
 
-    Long getCoins(Long userId);
+    int getCoins(Long userId);
 
-    Boolean addCoins(Long userId, Long coins);
+    Boolean addCoins(Long userId, int coins);
 
-    Boolean deductCoins(Long userId, Long deductedAmount);
+    Boolean deductCoins(Long userId, int deductedAmount);
 
-    Boolean validateBalance(Long userId, Long amount);
-
+    Boolean validateBalance(Long userId, int amount);
 
 }
